task 14

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -6,14 +6,14 @@
  */
 int main(void)
 {
-	int n1, n2, nextNum, i;
+	unsigned long long n1, n2, nextNum, i;
 
 	n1 = 1, n2 = 2;
-	printf("%d, %d, ", n1, n2);
+	printf("%llu, %llu, ", n1, n2);
 	for (i = 3; i <= 50; i++)
 	{
 		nextNum = n1 + n2;
-		printf("%d", nextNum);
+		printf("%llu", nextNum);
 		if (i != 50)
 		{
 			printf(", ");
